Make item replacement mixin more flexible
Avoid leading the user to use hard coded values.
Instead, the values of the preexisting item can be used when recreating the
item.
This also avoids forcing the user to lookup the item's namespaceID and
numeric ID and also hopefully makes the code more future-proof in case an
item's ID ever changes for any reason.